backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.3.1",
    "fastapi[standard]>=0.128.0",
    "flower>=2.0.1",
    "orjson>=3.11.5",
    "phonenumbers>=9.0.21",
    "prometheus-client>=0.23.1",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "structlog>=25.5.0",
    "user-agents>=2.2.0",
    "uuid6>=2025.0.1",
    "uvicorn[standard]>=0.40.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"
exclude = [".git", "__pycache__", ".venv", ".vscode"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "W191"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
